refactor: refactor getSpectralRules tool
