+++ src/extension.ts
@@ -47,9 +47,9 @@ import { testInPostman } from './commands/testInPostman';
 import { ErrorProperties, TelemetryProperties } from './common/telemetryEvent';
 import { LearnMoreAboutAPICatalog, doubleClickDebounceDelay, selectedNodeKey } from './constants';
 import { getPlugins } from './copilot/getPlugins';
-import { GetSpectralRulesTool } from './copilot/getSpectralRulesTool';
 import { ExportApiSpecificationTool } from './copilot/tools/exportApiSpecification';
 import { GetCredentialTool } from './copilot/tools/getCredential';
+import { GetSpectralRulesTool } from './copilot/tools/getSpectralRulesTool';
 import { SearchApiDefinitionsTool } from './copilot/tools/searchApiDefinitions';
 import { SearchApiDeploymentsTool } from './copilot/tools/searchApiDeployments';
 import { SearchApisTool } from './copilot/tools/searchApis';
